cc: codegen: move expression visitors to separate file
